replace constant fraction
