+++ lib/PolynomialTermOperations.js
@@ -3,7 +3,6 @@
 // Operations on polynomial term nodes
 
 const clone = require('./clone');
-const ConstantFraction = require('./ConstantFraction');
 const evaluateArithmetic = require('./evaluateArithmetic');
 const evaluateConstantSum = require('./evaluateConstantSum');
 const MathChangeTypes = require('./MathChangeTypes');
