+++ lib/evaluateConstantSum.js
@@ -1,7 +1,8 @@
 'use strict';
 
+const addConstantFractions = require('./simplifyFractions/addConstantFractions');
+const addConstantAndFraction = require('./simplifyFractions/addConstantAndFraction');
 const clone = require('./clone');
-const ConstantFraction = require('./ConstantFraction');
 const evaluateArithmetic = require('./evaluateArithmetic');
 const MathChangeTypes = require('./MathChangeTypes');
 const NodeCreator = require('./NodeCreator');
@@ -25,8 +26,8 @@ function evaluateConstantSum(node) {
   // functions needed to evaluate the sum
   const summingFunctions = [
     evaluateArithmetic,
-    ConstantFraction.addConstantFractions,
-    ConstantFraction.addConstantAndFraction,
+    addConstantFractions,
+    addConstantAndFraction,
   ];
   for (let i = 0; i < summingFunctions.length; i++) {
     const status = summingFunctions[i](node);
@@ -63,7 +64,7 @@ function evaluateConstantSum(node) {
   // (which is the case if it's a list surrounded by parenthesis)
   if (NodeType.isParenthesis(fractions)) {
     const fractionList = fractions.content;
-    const evaluateStatus = ConstantFraction.addConstantFractions(fractionList);
+    const evaluateStatus = addConstantFractions(fractionList);
     status = NodeStatus.childChanged(newNode, evaluateStatus, 1);
     substeps.push(status);
     newNode = NodeStatus.resetChangeGroups(status.newNode);
