+++ lib/simplifyFractions/divideByGCD.js
@@ -1,0 +1,60 @@
+'use strict';
+
+const math = require('mathjs');
+const MathChangeTypes = require('../MathChangeTypes');
+const NodeCreator = require('../NodeCreator');
+const NodeStatus = require('../NodeStatus');
+const NodeType = require('../NodeType');
+
+// Simplifies a fraction (with constant numerator and denominator) by dividing
+// the top and bottom by the GCD, if possible.
+// e.g. 2/4 --> 1/2    10/5 --> 2x
+// Also simplified negative signs
+// e.g. -1/-3 --> 1/3   4/-5 --> -4/5
+// Note that -4/5 doesn't need to be simplified.
+// Note that our goal is for the denominator to always be positive. If it
+// isn't, we can simplify signs.
+// Returns a NodeStatus object
+function divideByGCD(fraction) {
+  if (!NodeType.isOperator(fraction) || fraction.op !== '/') {
+    return NodeStatus.noChange(fraction);
+  }
+  // If it's not an integer fraction, all we can do is simplify signs
+  if (!NodeType.isIntegerFraction(fraction, true)) {
+    return NodeStatus.noChange(fraction);
+  }
+
+  const numeratorValue = parseInt(fraction.args[0].eval());
+  const denominatorValue = parseInt(fraction.args[1].eval());
+
+  // The gcd is what we're dividing the numerator and denominator by.
+  let gcd = math.gcd(numeratorValue, denominatorValue);
+  // A greatest common denominator is technically defined as always positive,
+  // but since our goal is to reduce negative signs or move them to the
+  // numerator, a negative denominator always means we want to flip signs
+  // of both numerator and denominator.
+  // e.g. -1/-3 --> 1/3   4/-5 --> -4/5
+  if (denominatorValue < 0) {
+    gcd *= -1;
+  }
+
+  if (gcd === 1) {
+    return NodeStatus.noChange(fraction);
+  }
+
+  const newNumeratorNode = NodeCreator.constant(numeratorValue/gcd);
+  const newDenominatorNode = NodeCreator.constant(denominatorValue/gcd);
+  let newFraction;
+  if (parseFloat(newDenominatorNode.value) === 1) {
+    newFraction = newNumeratorNode;
+  }
+  else {
+    newFraction = NodeCreator.operator(
+      '/', [newNumeratorNode, newDenominatorNode]);
+  }
+
+  return NodeStatus.nodeChanged(
+    MathChangeTypes.SIMPLIFY_FRACTION, fraction, newFraction);
+}
+
+module.exports = divideByGCD;
