+++ lib/simplifyFractions/simplifyFractions.js
@@ -2,25 +2,34 @@
 
 /*
  * Performs simpifications on fractions: adding and cancelling out.
+ *
+ * Note: division is represented in mathjs as an operator node with op '/'
+ * and two args, where arg[0] is the numerator and arg[1] is the denominator
+
+// This module manipulates fractions with constants in the numerator and
+// denominator. For more complex/general fractions, see Fraction.js
+
  */
 
+const addConstantFractions = require('./addConstantFractions');
+const addConstantAndFraction = require('./addConstantAndFraction');
 const cancelLikeTerms = require('./cancelLikeTerms');
-const ConstantFraction = require('../ConstantFraction');
+const divideByGCD = require('./divideByGCD');
 const simplifyFractionSigns = require('./simplifyFractionSigns');
 const simplifyPolynomialFraction = require('./simplifyPolynomialFraction');
+
 const NodeStatus = require('../NodeStatus');
-const PolynomialTermOperations = require('../PolynomialTermOperations');
 const TreeSearch = require('../TreeSearch');
 
 const SIMPLIFICATION_FUNCTIONS = [
   // e.g. 2/3 + 5/6
-  ConstantFraction.addConstantFractions,
+  addConstantFractions,
   // e.g. 4 + 5/6 or 4.5 + 6/8
-  ConstantFraction.addConstantAndFraction,
+  addConstantAndFraction,
   // e.g. 2/-9  ->  -2/9      e.g. -2/-9  ->  2/9
   simplifyFractionSigns,
   // e.g. 8/12  ->  2/3 (divide by GCD 4)
-  ConstantFraction.divideByGCD,
+  divideByGCD,
   // e.g. 2x/4 -> x/2 (divideByGCD but for coefficients of polynomial terms)
   simplifyPolynomialFraction,
   // e.g. (2x * 5) / 2x  ->  5
