+++ lib/simplifyFractions/simplifyPolynomialFraction.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const clone = require('../clone');
-const ConstantFraction = require('../ConstantFraction');
+const divideByGCD = require('./divideByGCD');
 const evaluateArithmetic = require('../evaluateArithmetic');
 const NodeCreator = require('../NodeCreator');
 const NodeStatus = require('../NodeStatus');
@@ -23,7 +23,7 @@ function simplifyPolynomialFraction(node) {
   }
 
   const coefficientSimplifications = [
-    ConstantFraction.divideByGCD, // for integer fractions
+    divideByGCD, // for integer fractions
     evaluateArithmetic, // for decimal fractions
   ];
 
@@ -43,6 +43,6 @@ function simplifyPolynomialFraction(node) {
   }
 
   return NodeStatus.noChange(node);
-};
+}
 
 module.exports = simplifyPolynomialFraction;
